refactor: Module schema_generator renamed to main. Some type hints improvements
