+++ dataclasses_avroschema/__init__.py
@@ -1,64 +1,64 @@
-from .schema_generator import AvroModel  # noqa: I001
-from .types import Int32, Float32, TimeMicro, DateTimeMicro, condecimal, confixed
-from .model_generator.generator import BaseClassEnum, ModelType, ModelGenerator
 from .fields.field_utils import (
-    BOOLEAN,
-    NULL,
-    INT,
-    FLOAT,
-    LONG,
-    DOUBLE,
-    BYTES,
-    STRING,
     ARRAY,
-    ENUM,
-    MAP,
-    FIXED,
+    BOOLEAN,
+    BYTES,
     DATE,
-    UUID,
     DECIMAL,
-    RECORD,
-    TIME_MILLIS,
-    TIME_MICROS,
-    TIMESTAMP_MILLIS,
-    TIMESTAMP_MICROS,
+    DOUBLE,
+    ENUM,
+    FIXED,
+    FLOAT,
+    INT,
     LOGICAL_DATE,
-    LOGICAL_TIME_MILIS,
-    LOGICAL_TIME_MICROS,
-    LOGICAL_DATETIME_MILIS,
     LOGICAL_DATETIME_MICROS,
+    LOGICAL_DATETIME_MILIS,
+    LOGICAL_TIME_MICROS,
+    LOGICAL_TIME_MILIS,
     LOGICAL_UUID,
+    LONG,
+    MAP,
+    NULL,
     PYTHON_TYPE_TO_AVRO,
+    RECORD,
+    STRING,
+    TIME_MICROS,
+    TIME_MILLIS,
+    TIMESTAMP_MICROS,
+    TIMESTAMP_MILLIS,
+    UUID,
 )
 from .fields.fields import (
-    ImmutableField,
-    StringField,
-    IntField,
-    LongField,
+    AvroField,
     BooleanField,
-    DoubleField,
-    FloatField,
     BytesField,
-    NoneField,
     ContainerField,
-    ListField,
-    TupleField,
-    DictField,
-    UnionField,
-    LiteralField,
-    FixedField,
-    EnumField,
-    SelfReferenceField,
     DateField,
     DatetimeField,
     DatetimeMicroField,
-    TimeMilliField,
-    TimeMicroField,
-    UUIDField,
     DecimalField,
+    DictField,
+    DoubleField,
+    EnumField,
+    FixedField,
+    FloatField,
+    ImmutableField,
+    IntField,
+    ListField,
+    LiteralField,
+    LongField,
+    NoneField,
     RecordField,
-    AvroField,
+    SelfReferenceField,
+    StringField,
+    TimeMicroField,
+    TimeMilliField,
+    TupleField,
+    UnionField,
+    UUIDField,
 )
+from .main import AVRO, AVRO_JSON, AvroModel
+from .model_generator.generator import BaseClassEnum, ModelGenerator, ModelType
+from .types import DateTimeMicro, Float32, Int32, TimeMicro, condecimal, confixed
 
 __all__ = [
     "AvroModel",
@@ -125,4 +125,6 @@ __all__ = [
     "DecimalField",
     "RecordField",
     "AvroField",
+    "AVRO",
+    "AVRO_JSON",
 ]
