+++ dataclasses_avroschema/pydantic/v1/main.py
@@ -2,8 +2,7 @@ from typing import Any, Dict, Type, TypeVar
 
 from fastavro.validation import validate
 
-from dataclasses_avroschema import serialization
-from dataclasses_avroschema.schema_generator import AVRO, AvroModel
+from dataclasses_avroschema import AVRO, AvroModel, serialization
 from dataclasses_avroschema.types import JsonDict
 from dataclasses_avroschema.utils import standardize_custom_type
